+++ src/components/components/Tweet.tsx
@@ -1,48 +1,16 @@
 import { TwitterTweetEmbed } from "react-twitter-embed";
 import { TwitterRepository } from "../../repository/TwitterRepository";
+import { createResource } from "../utils/helpers";
 const ACCOUNT_NAME = "ochakai_vrc";
-
-const wrapPromise = <T,>(
-  promise: Promise<T>
-): {
-  read(): T;
-} => {
-  let status = "pending";
-  let result: T;
-  const suspender = promise.then(
-    (r) => {
-      status = "success";
-      result = r;
-    },
-    (e) => {
-      status = "error";
-      result = e;
-    }
-  );
-  return {
-    read: () => {
-      console.log(status);
-      if (status === "pending") {
-        throw suspender;
-      } else if (status === "error") {
-        throw result;
-      } else if (status === "success") {
-        return result;
-      }
-      throw new Error("Unexpected status");
-    },
-  };
-};
 
 const useTweetData = (accountName: string) => {
   const repository = new TwitterRepository();
-  return wrapPromise(repository.getPinnedTweetId(accountName));
+  return createResource(repository.getPinnedTweetId(accountName));
 };
 
+const resource = useTweetData(ACCOUNT_NAME);
 const SuspenseTwitterTweetEmbed = (): JSX.Element => {
-  const pinnedTweetID = useTweetData(ACCOUNT_NAME).read();
-  const repository = new TwitterRepository();
-  console.log("getPinnedTweetId", repository.getPinnedTweetId(ACCOUNT_NAME));
+  const pinnedTweetID = resource.read();
   return <TwitterTweetEmbed tweetId={pinnedTweetID} />;
 };
